feat : WinLotto 매치 결과 계산 기능 구현 및 테스트

+++ src/main/java/lotto/domain/Lotto.java
@@ -36,8 +36,8 @@ public class Lotto {
         return lottoNumbers.contains(lottoNumber);
     }
 
-    public long match(final Lotto compareLotto) {
-        return lottoNumbers.stream()
+    public int match(final Lotto compareLotto) {
+        return (int) lottoNumbers.stream()
                 .filter(compareLotto::containNumber)
                 .count();
     }
